Reorganize extension methods + remove unused/obsolete.
