+++ ICSharpCode.Decompiler/CSharp/Syntax/PatternMatching/INode.cs
@@ -65,5 +65,30 @@ namespace ICSharpCode.Decompiler.CSharp.Syntax.PatternMatching
 				throw new ArgumentNullException("pattern");
 			return pattern.DoMatch(other, PatternMatching.Match.CreateNew());
 		}
+
+		public static AstType ToType(this Pattern pattern)
+		{
+			return pattern;
+		}
+
+		public static Expression ToExpression(this Pattern pattern)
+		{
+			return pattern;
+		}
+
+		public static Statement ToStatement(this Pattern pattern)
+		{
+			return pattern;
+		}
+
+		public static Expression WithName(this Expression node, string patternGroupName)
+		{
+			return new NamedNode(patternGroupName, node);
+		}
+
+		public static Statement WithName(this Statement node, string patternGroupName)
+		{
+			return new NamedNode(patternGroupName, node);
+		}
 	}
 }
