+++ ICSharpCode.Decompiler/CSharp/Syntax/SyntaxExtensions.cs
@@ -40,5 +40,40 @@ namespace ICSharpCode.Decompiler.CSharp.Syntax
 					return false;
 			}
 		}
+
+		/// <summary>
+		/// Returns true if <paramref name="operatorType"/> is bitwise and, bitwise or, or exclusive or.
+		/// </summary>
+		public static bool IsBitwise(this BinaryOperatorType operatorType)
+		{
+			return operatorType == BinaryOperatorType.BitwiseAnd
+				|| operatorType == BinaryOperatorType.BitwiseOr
+				|| operatorType == BinaryOperatorType.ExclusiveOr;
+		}
+
+		public static Statement GetNextStatement(this Statement statement)
+		{
+			AstNode next = statement.NextSibling;
+			while (next != null && !(next is Statement))
+				next = next.NextSibling;
+			return (Statement)next;
+		}
+
+		public static bool IsArgList(this AstType type)
+		{
+			var simpleType = type as SimpleType;
+			return simpleType != null && simpleType.Identifier == "__arglist";
+		}
+
+		public static void AddNamedArgument(this Syntax.Attribute attribute, string name, Expression argument)
+		{
+			attribute.Arguments.Add(new AssignmentExpression(new IdentifierExpression(name), argument));
+		}
+
+		public static T Detach<T>(this T node) where T : AstNode
+		{
+			node.Remove();
+			return node;
+		}
 	}
 }
